+++ packages/manager/modules/pci/src/projects/project/private-networks/add/add.component.js
@@ -5,7 +5,6 @@ export default {
   bindings: {
     goBack: '<',
     projectId: '<',
-    messageContainer: '<',
   },
   controller,
   template,